+++ src/controllers/UserController.ts
@@ -1,10 +1,13 @@
+import { isArray } from "lodash";
 import { ObjectId } from "mongodb";
 import { Body, Delete, Get, Patch, Post, Queries, Route, Security, Tags } from "tsoa/dist";
 
 import BaseController from "@/controllers/BaseController";
-import type { User } from "@/entities";
+import type { Role, User } from "@/entities";
 import type { HiddenBodyKeys, ResponseData } from "@/interfaces";
-import { IDeleteQueryParams, IGetQueryParams, IPostQueryParams } from "@/interfaces";
+import { IDeleteQueryParams, IGetQueryParams, IPostQueryParams, respondFailure, respondSuccess } from "@/interfaces";
+import { DB } from "@/modules/api/DB";
+import { filterRole } from "@/plugins/user-utils";
 import UserService from "@/services/UserService";
 import WorkspaceService from "@/services/WorkspaceService";
 
@@ -16,6 +19,8 @@ interface JoinWorkspaceBody {
 @Tags("User")
 @Route("user")
 export default class UserController extends BaseController<User> {
+	service: UserService;
+
 	constructor() {
 		super(new UserService());
 	}
@@ -23,13 +28,18 @@ export default class UserController extends BaseController<User> {
 	@Security("api_key")
 	@Security("jwt")
 	@Get("/")
-	read(@Queries() queryParams?: IGetQueryParams) {
-		// console.log("this.user.activeWorkspace :>> ", this.user.activeWorkspace);
-		this.filter.workspaces = this.workspace._id;
-		return super.read();
+	async read(@Queries() queryParams?: IGetQueryParams) {
+		const res = await super.read();
+
+		console.log("[1] res.data :>> ", res.data);
+		res.data = isArray(res.data)
+			? await filterRole(this.workspace._id.toString(), res.data)
+			: await filterRole(this.workspace._id.toString(), [res.data]);
+		console.log("[2] res.data :>> ", res.data);
+		return res;
 	}
 
-	@Security("api_key")
+	@Security("api_key2")
 	@Security("jwt")
 	@Post("/")
 	create(@Body() body: Omit<User, keyof HiddenBodyKeys>, @Queries() queryParams?: IPostQueryParams) {
@@ -39,7 +49,21 @@ export default class UserController extends BaseController<User> {
 	@Security("api_key")
 	@Security("jwt")
 	@Patch("/")
-	update(@Body() body: Omit<User, keyof HiddenBodyKeys>, @Queries() queryParams?: IPostQueryParams) {
+	async update(@Body() body: Omit<User, keyof HiddenBodyKeys>, @Queries() queryParams?: IPostQueryParams) {
+		if (body.roles) {
+			const roleId = body.roles;
+			const oldRoles = this.user.roles.filter((role) => (role as Role).workspace.toString() === this.workspace._id.toString());
+
+			const newRole = await DB.findOne<Role>("role", { _id: roleId });
+			const newRoles = [
+				...this.user.roles.filter((role) => !oldRoles.map((r) => (r as Role)._id.toString()).includes((role as Role)._id.toString())),
+				newRole,
+			];
+
+			const newRoleIds = newRoles.map((role) => (role as Role)._id);
+
+			body.roles = newRoleIds;
+		}
 		return super.update(body);
 	}
 
@@ -48,6 +72,35 @@ export default class UserController extends BaseController<User> {
 	@Delete("/")
 	delete(@Queries() queryParams?: IDeleteQueryParams) {
 		return super.delete();
+	}
+
+	@Security("api_key")
+	@Security("jwt")
+	@Patch("/assign-role")
+	async assignRole(@Body() data: { roleId: string }) {
+		if (!data.roleId) return respondFailure({ msg: `Role ID is required.` });
+		if (!this.user) return respondFailure({ msg: `User not found.` });
+
+		const { roleId } = data;
+		const currentRoles = this.user.roles.filter((role) => (role as Role).workspace.toString() === this.workspace._id.toString());
+		const prevRole = currentRoles[0] as Role;
+
+		const newRole = await DB.findOne<Role>("role", { _id: roleId });
+		const newRoles = [...this.user.roles.filter((role) => (role as Role)._id.toString() !== prevRole._id.toString()), newRole];
+		const newRoleIds = newRoles.map((role) => (role as Role)._id);
+
+		let [updatedUser] = await DB.update<User>(
+			"user",
+			{ _id: this.user._id },
+			{ roles: newRoleIds },
+			{ populate: ["roles", "workspaces", "activeWorkspace"] }
+		);
+		if (!updatedUser) return respondFailure({ msg: `Failed to assign "${newRole.name}" role to "${this.user.slug}" user.` });
+
+		// filter roles & workspaces before returning
+		[updatedUser] = await filterRole(this.workspace._id.toString(), [updatedUser]);
+
+		return respondSuccess({ data: updatedUser });
 	}
 
 	@Security("api_key")
@@ -81,7 +134,8 @@ export default class UserController extends BaseController<User> {
 
 			let updatedUser = [user];
 
-			const isUserJoinedThisWorkspace = (user.workspaces || []).map((id) => id.toString()).includes(wsId);
+			const workspaceIds = user.workspaces || [];
+			const isUserJoinedThisWorkspace = workspaceIds.map((id) => id.toString()).includes(wsId);
 			// console.log("isUserJoinedThisWorkspace :>> ", isUserJoinedThisWorkspace);
 
 			const isWorkspaceActive = typeof user.activeWorkspace !== "undefined" && user.activeWorkspace.toString() === wsId;
@@ -89,7 +143,7 @@ export default class UserController extends BaseController<User> {
 
 			// console.log("user.workspaces :>> ", user.workspaces);
 			if (!isUserJoinedThisWorkspace) {
-				updatedUser = await this.service.update({ _id: userId }, { $push: { workspaces: workspace._id } }, { raw: true });
+				updatedUser = await this.service.update({ _id: userId }, { workspaces: [...workspaceIds, wsId] });
 			}
 			// console.log("[1] updatedUser :>> ", updatedUser[0]);