+++ src/modules/passports/jwtStrategy.ts
@@ -9,7 +9,7 @@ import type { VerifiedCallback } from "passport-jwt";
 import { ExtractJwt, Strategy } from "passport-jwt";
 
 import { Config } from "@/app.config";
-import type { AccessTokenInfo, Role, User } from "@/entities";
+import type { AccessTokenInfo, Role, User, Workspace } from "@/entities";
 import type ServiceAccount from "@/entities/ServiceAccount";
 
 import { DB } from "../api/DB";
@@ -134,7 +134,8 @@ export const jwtStrategy = new Strategy(
 			if (isEmpty(user.workspaces)) {
 				updateData.workspaces = [workspaceId];
 			} else {
-				if (!user.workspaces.includes(workspaceId)) updateData.workspaces = [...user.workspaces, workspaceId];
+				if (!user.workspaces.includes(workspaceId))
+					updateData.workspaces = [...(user.workspaces || []).map((ws) => (ws as Workspace)._id), workspaceId];
 			}
 
 			// set default roles if this user doesn't have one
@@ -148,6 +149,9 @@ export const jwtStrategy = new Strategy(
 				populate: ["roles", "workspaces", "activeWorkspace"],
 			});
 
+			// filter roles
+			// [user] = await filterRole(user.activeWorkspace.toString(), [user]);
+
 			return done(null, user);
 		}
 
@@ -157,6 +161,9 @@ export const jwtStrategy = new Strategy(
 			{ _id: new ObjectId(payload.id) },
 			{ populate: ["roles", "workspaces", "activeWorkspace"] }
 		);
+
+		// filter roles
+		// if (user) [user] = await filterRole(user.activeWorkspace.toString(), [user]);
 
 		// console.log(`[3] jwtStrategy > ServiceAccount :>> `, user.name, user._id);