+++ src/routes/api/v1/git-provider.ts
@@ -4,6 +4,7 @@ import GitProviderController from "@/controllers/GitProviderController";
 import { authenticate } from "@/middlewares/authenticate";
 import { authorize } from "@/middlewares/authorize";
 import { processApiRequest } from "@/middlewares/process-api-request";
+import { registerController } from "@/middlewares/register-controller";
 
 const router = express.Router();
 
@@ -11,6 +12,7 @@ const controller = new GitProviderController();
 
 router
 	.use(authenticate, authorize)
+	.use(registerController(controller))
 	.use(controller.parsePagination.bind(controller))
 	.use(controller.parseFilter.bind(controller))
 	.use(controller.parseBody.bind(controller))